requirements.txt
pyinstaller
boto3
python-dotenv